product editing data loading

+++ app/src/main/java/com/kolshop/kolshopmaterial/model/realm/Product.java
@@ -20,12 +20,12 @@ public class Product extends RealmObject{
     private String name;
     private String description;
     private String brand;
-    private int brandId;
-    private int userId;
-    private int productCategoryId;
+    private Long brandId;
+    private Long userId;
+    private Long productCategoryId;
     private RealmList<ProductVariety> listProductVariety;
 
-    public Product(String id,String name, String description, String brand, int brandId, int userId, int productCategoryId) {
+    public Product(String id,String name, String description, String brand, Long brandId, Long userId, Long productCategoryId) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -40,27 +40,27 @@ public class Product extends RealmObject{
 
     }
 
-    public int getProductCategoryId() {
+    public Long getProductCategoryId() {
         return productCategoryId;
     }
 
-    public void setProductCategoryId(int productCategoryId) {
+    public void setProductCategoryId(Long productCategoryId) {
         this.productCategoryId = productCategoryId;
     }
 
-    public int getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
-    public int getBrandId() {
+    public Long getBrandId() {
         return brandId;
     }
 
-    public void setBrandId(int brandId) {
+    public void setBrandId(Long brandId) {
         this.brandId = brandId;
     }
